Revert "Improve UX for COSMIC"
This reverts commit 239420148a583c45d9df08de27be75cc537dc62b.

Please Stop Theming My App should be focused on distros or packagers theming the apps, shipping them broken, it's not user's fault as they don't know what is going on.
popOS is one of the most used distros, and most of their users are new users that saw them in YouTube, please don't break their experience.