Fix: Import never finishes (#301)
* Add separate callback to update progressbar

* Black formatting

* Make watchdog useful

The app was still not thread safe because the dialog was
closed from a thread which was the likely cause of the
crashing.

Still needs more testing to validate this solution

* Prevent opening preferences while importing

* Formatting changes & regrouping of methods

- Rearranged methods into groups based on public or private and their
usage. No functional changes, just reorganizing for convenience.

 - Renamed import_callback -> finish_import.
 - Renamed __watchdog -> monitor_import

If you don't like it, this can be reverted.

* Black formatting

* Cleanup for review

* Fix attempt 2 & remove public/private comments

Readded watchdog in its original form.

Removed comments for public or private methods. Comments for import
actions/gui are still there